rename existing components
