+++ addon/components/em-select.js
@@ -1,5 +1,5 @@
 import Ember from 'ember';
-import layout from '../templates/components/mc-select';
+import layout from '../templates/components/em-select';
 
 export default Ember.Component.extend({
     layout,
