+++ addon/templates/components/em-card.hbs
