+++ addon/templates/components/em-copyright.hbs
