+++ app/components/em-card.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-materializecss/components/em-card';
