+++ app/components/em-copyright.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-materializecss/components/em-copyright';
