+++ app/components/em-input.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-materializecss/components/em-input';
