+++ app/components/em-select.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-materializecss/components/em-select';
