+++ package.json
@@ -14,7 +14,7 @@
   "repository": "",
   "scripts": {
     "build": "ember build",
-    "start": "ember server",
+    "start": "rm -rf dist tmp && ember server",
     "test": "ember try:each"
   },
   "dependencies": {
