+++ tests/dummy/app/templates/cards.hbs
@@ -1,17 +1,17 @@
-{{#mc-card title="Cards"}}
+{{#em-card title="Cards"}}
 
   <div class="section">
     <p>A basic card.</p>
-    {{#mc-card title="Basic Card"}}
+    {{#em-card title="Basic Card"}}
       <p>Some content here.</p>
-    {{/mc-card}}
+    {{/em-card}}
   </div>
 
   <div class="section">
     <p>A card with some color</p>
-    {{#mc-card title="I like teal" class="teal lighten-2"}}
+    {{#em-card title="I like teal" class="teal lighten-2"}}
       <p class="white-text">A card of any color</p>
-    {{/mc-card}}
+    {{/em-card}}
   </div>
 
-{{/mc-card}}
+{{/em-card}}
