+++ tests/dummy/app/templates/components/app-layout.hbs
@@ -1,7 +1,7 @@
 <header>
   <nav>
     <div class="nav-wrapper">
-      {{link-to "Ember MaterializeCss" "index" class="brand-logo"}}
+      {{link-to "Ember Materialize" "index" class="brand-logo"}}
       <ul class="right">
         {{#link-to "forms" tagName="li"}}
           {{link-to "Forms" "forms"}}
@@ -24,7 +24,7 @@
 <footer>
   <footer class="page-footer">
     <div class="footer-copyright">
-      {{mc-copyright startYear=2016}}
+      {{em-copyright startYear=2016}}
     </div>
   </footer>
 </footer>
