+++ tests/dummy/app/templates/copyright.hbs
@@ -1,4 +1,4 @@
-{{#mc-card title="Copyright"}}
+{{#em-card title="Copyright"}}
   <p>A basic copyright component that never needs updating.</p>
-  {{mc-copyright startYear=1974 text="Me, all rights reserved"}}
-{{/mc-card}}
+  {{em-copyright startYear=1974 text="Me, all rights reserved"}}
+{{/em-card}}
