+++ tests/dummy/app/templates/forms.hbs
@@ -1,40 +1,40 @@
-{{#mc-card title="Forms"}}
+{{#em-card title="Forms"}}
 
   <div class="section">
     <h5 class="title">Inputs</h5>
 
     <div class="section">
       <p>A basic text input with a label</p>
-      {{mc-input label="Full Name"}}
+      {{em-input label="Full Name"}}
     </div>
 
     <div class="section">
       <p>A text input with an icon</p>
-      {{mc-input label="Email" icon="email"}}
+      {{em-input label="Email" icon="email"}}
     </div>
 
     <div class="section">
       <p>A text input with a placeholder</p>
-      {{mc-input placeholder="Your name here"}}
+      {{em-input placeholder="Your name here"}}
     </div>
 
     <div class="section">
       <p>A disabled text input</p>
-      {{mc-input label="Name" disabled=true}}
+      {{em-input label="Name" disabled=true}}
     </div>
 
     <div class="section">
       <p>A password input</p>
-      {{mc-input label="Password" type="password"}}
+      {{em-input label="Password" type="password"}}
     </div>
 
     <h5>Selects</h5>
 
     <div class="section">
       <p>A basic select input</p>
-      {{mc-select label="Your Options" content=selectOptions optionLabelPath="value" optionValuePath="value"}}
+      {{em-select label="Your Options" content=selectOptions optionLabelPath="value" optionValuePath="value"}}
     </div>
 
   </div>
 
-{{/mc-card}}
+{{/em-card}}
