+++ tests/dummy/app/templates/index.hbs
@@ -1,12 +1,9 @@
-{{#mc-card title="Welcome"}}
+{{#em-card title="Welcome"}}
   <p>
-    Welcome to Ember Materialize CSS.
+    Welcome to Ember Materialize.
   </p>
   <p>
-    This Ember Addon does not aim to be a replacement for "ember-cli-materialize".
+    This Ember Addon aims to be an alternative for developer's looking for a simple wrapper for the Materialize CSS project.
   </p>
-  <p>
-    Having struggled with external dependencies during upgrades, it seemed worthwhile to provide an addon that installs only Materialize assets.
-  </p>
-{{/mc-card}}
+{{/em-card}}
